POLISCI 792L Manosphere Final Paper
The file attached is the final paper I turned in for my last graduate course in Political Science, which was "Law and Democracy." In this paper, I explored the pipeline of the manosphere to the alt-right. This paper was a research design paper, so I ended up using computer generated results to run a survey that I would've have liked to do for this research paper.